Add CacheTypes enum and VerifierOptions class, and remove ProcessOptions and RabbitMqOptions classes
