+++ src/ProjectOrigin.Registry.Server/Options/CacheOptions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace ProjectOrigin.Registry.Server.Models;
+namespace ProjectOrigin.Registry.Server.Options;
 
 public record CacheOptions : IValidatableObject
 {
