+++ src/ProjectOrigin.Registry.Server/Options/CacheTypes.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace ProjectOrigin.Registry.Server.Models;
+namespace ProjectOrigin.Registry.Server.Options;
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
 public enum CacheTypes { InMemory, Redis }
