+++ src/ProjectOrigin.Registry.Server/Options/RedisOptions.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace ProjectOrigin.Registry.Server.Models;
+namespace ProjectOrigin.Registry.Server.Options;
 
 public record RedisOptions : IValidatableObject
 {
