+++ src/ProjectOrigin.Registry.Server/Services/ConsistentHashRingQueueResolver.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Options;
 using Murmur;
 using ProjectOrigin.Registry.Server.Interfaces;
 using ProjectOrigin.Registry.Server.Models;
+using ProjectOrigin.Registry.Server.Options;
 
 namespace ProjectOrigin.Registry.Server.Services;
